requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
asyncio==3.4.3
attrs==25.1.0
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.7.0
Django==5.1.4
django-ckeditor-5==0.2.17
django-debug-toolbar==5.0.1
django-js-asset==3.0.1
django-mptt==0.16.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
djangorestframework==3.16.1
flower==2.0.1
fonttools==4.56.0
frozenlist==1.5.0
humanize==4.14.0
idna==3.10
kombu==5.5.4
multidict==6.1.0
numpy==2.2.3
packaging==25.0
pillow==11.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.1
rabbitmq==0.2.0
requests==2.32.3
six==1.17.0
slugify==0.0.1
sqlparse==0.5.3
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.2
transliterate==1.10.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.3.0
uWSGI==2.0.31
vine==5.1.0
wcwidth==0.2.14
weasyprint==64.1
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
